增加fastNLP.embeddings模块并修改对应的现有代码以适配fastNLP.embeddings
