+++ fastNLP/io/data_loader/matching.py
@@ -6,7 +6,7 @@ from ...core.const import Const
 from ...core.vocabulary import Vocabulary
 from ..base_loader import DataBundle, DataSetLoader
 from ..file_utils import _get_base_url, cached_path, PRETRAINED_BERT_MODEL_DIR
-from ...modules.encoder._bert import BertTokenizer
+from ...modules.encoder.bert import BertTokenizer
 
 
 class MatchingLoader(DataSetLoader):
