+++ fastNLP/modules/encoder/__init__.py
@@ -1,18 +1,10 @@
 __all__ = [
-    # "BertModel",
+    "BertModel",
     
     "ConvolutionCharEncoder",
     "LSTMCharEncoder",
     
     "ConvMaxpool",
-    
-    "Embedding",
-    "StaticEmbedding",
-    "ElmoEmbedding",
-    "BertEmbedding",
-    "StackEmbedding",
-    "LSTMCharEmbedding",
-    "CNNCharEmbedding",
     
     "LSTM",
     
@@ -31,12 +23,10 @@ __all__ = [
 
     "MultiHeadAttention",
 ]
-from ._bert import BertModel
-from .bert import BertWordPieceEncoder
+
+from .bert import BertModel
 from .char_encoder import ConvolutionCharEncoder, LSTMCharEncoder
 from .conv_maxpool import ConvMaxpool
-from .embedding import Embedding, StaticEmbedding, ElmoEmbedding, BertEmbedding, \
-    StackEmbedding, LSTMCharEmbedding, CNNCharEmbedding
 from .lstm import LSTM
 from .star_transformer import StarTransformer
 from .transformer import TransformerEncoder
