+++ fastNLP/modules/encoder/_elmo.py
@@ -4,18 +4,13 @@
 
 from typing import Optional, Tuple, List, Callable
 
-import os
-
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn.utils.rnn import PackedSequence, pad_packed_sequence
-from ...core.vocabulary import Vocabulary
-import json
-import pickle
 
 from ..utils import get_dropout_mask
-import codecs
+
 
 class LstmCellWithProjection(torch.nn.Module):
     """
@@ -541,188 +536,3 @@ class Highway(torch.nn.Module):
             gate = torch.sigmoid(gate)
             current_input = gate * linear_part + (1 - gate) * nonlinear_part
         return current_input
-
-
-class _ElmoModel(nn.Module):
-    """
-    该Module是ElmoEmbedding中进行所有的heavy lifting的地方。做的工作，包括
-        (1) 根据配置，加载模型;
-        (2) 根据vocab，对模型中的embedding进行调整. 并将其正确初始化
-        (3) 保存一个words与chars的对应转换，获取时自动进行相应的转换
-        (4) 设计一个保存token的embedding，允许缓存word的表示。
-
-    """
-
-    def __init__(self, model_dir: str, vocab: Vocabulary = None, cache_word_reprs: bool = False):
-        super(_ElmoModel, self).__init__()
-        self.model_dir = model_dir
-        dir = os.walk(self.model_dir)
-        config_file = None
-        weight_file = None
-        config_count = 0
-        weight_count = 0
-        for path, dir_list, file_list in dir:
-            for file_name in file_list:
-                if file_name.__contains__(".json"):
-                    config_file = file_name
-                    config_count += 1
-                elif file_name.__contains__(".pkl"):
-                    weight_file = file_name
-                    weight_count += 1
-        if config_count > 1 or weight_count > 1:
-            raise Exception(f"Multiple config files(*.json) or weight files(*.hdf5) detected in {model_dir}.")
-        elif config_count == 0 or weight_count == 0:
-            raise Exception(f"No config file or weight file found in {model_dir}")
-
-        config = json.load(open(os.path.join(model_dir, config_file), 'r'))
-        self.weight_file = os.path.join(model_dir, weight_file)
-        self.config = config
-
-        OOV_TAG = '<oov>'
-        PAD_TAG = '<pad>'
-        BOS_TAG = '<bos>'
-        EOS_TAG = '<eos>'
-        BOW_TAG = '<bow>'
-        EOW_TAG = '<eow>'
-
-        # For the model trained with character-based word encoder.
-        char_lexicon = {}
-        with codecs.open(os.path.join(model_dir, 'char.dic'), 'r', encoding='utf-8') as fpi:
-            for line in fpi:
-                tokens = line.strip().split('\t')
-                if len(tokens) == 1:
-                    tokens.insert(0, '\u3000')
-                token, i = tokens
-                char_lexicon[token] = int(i)
-
-        # 做一些sanity check
-        for special_word in [PAD_TAG, OOV_TAG, BOW_TAG, EOW_TAG]:
-            assert special_word in char_lexicon, f"{special_word} not found in char.dic."
-
-        # 从vocab中构建char_vocab
-        char_vocab = Vocabulary(unknown=OOV_TAG, padding=PAD_TAG)
-        # 需要保证<bow>与<eow>在里面
-        char_vocab.add_word_lst([BOW_TAG, EOW_TAG, BOS_TAG, EOS_TAG])
-
-        for word, index in vocab:
-            char_vocab.add_word_lst(list(word))
-
-        self.bos_index, self.eos_index, self._pad_index = len(vocab), len(vocab) + 1, vocab.padding_idx
-        # 根据char_lexicon调整, 多设置一位，是预留给word padding的(该位置的char表示为全0表示)
-        char_emb_layer = nn.Embedding(len(char_vocab) + 1, int(config['char_cnn']['embedding']['dim']),
-                                      padding_idx=len(char_vocab))
-
-        # 读入预训练权重 这里的elmo_model 包含char_cnn和 lstm 的 state_dict
-        elmo_model = torch.load(os.path.join(self.model_dir, weight_file), map_location='cpu')
-
-        char_embed_weights = elmo_model["char_cnn"]['char_emb_layer.weight']
-
-        found_char_count = 0
-        for char, index in char_vocab:  # 调整character embedding
-            if char in char_lexicon:
-                index_in_pre = char_lexicon.get(char)
-                found_char_count += 1
-            else:
-                index_in_pre = char_lexicon[OOV_TAG]
-            char_emb_layer.weight.data[index] = char_embed_weights[index_in_pre]
-
-        print(f"{found_char_count} out of {len(char_vocab)} characters were found in pretrained elmo embedding.")
-        # 生成words到chars的映射
-        max_chars = config['char_cnn']['max_characters_per_token']
-
-        self.words_to_chars_embedding = nn.Parameter(torch.full((len(vocab) + 2, max_chars),
-                                                                fill_value=len(char_vocab),
-                                                                dtype=torch.long),
-                                                     requires_grad=False)
-        for word, index in list(iter(vocab)) + [(BOS_TAG, len(vocab)), (EOS_TAG, len(vocab) + 1)]:
-            if len(word) + 2 > max_chars:
-                word = word[:max_chars - 2]
-            if index == self._pad_index:
-                continue
-            elif word == BOS_TAG or word == EOS_TAG:
-                char_ids = [char_vocab.to_index(BOW_TAG)] + [char_vocab.to_index(word)] + [
-                    char_vocab.to_index(EOW_TAG)]
-                char_ids += [char_vocab.to_index(PAD_TAG)] * (max_chars - len(char_ids))
-            else:
-                char_ids = [char_vocab.to_index(BOW_TAG)] + [char_vocab.to_index(c) for c in word] + [
-                    char_vocab.to_index(EOW_TAG)]
-                char_ids += [char_vocab.to_index(PAD_TAG)] * (max_chars - len(char_ids))
-            self.words_to_chars_embedding[index] = torch.LongTensor(char_ids)
-
-        self.char_vocab = char_vocab
-
-        self.token_embedder = ConvTokenEmbedder(
-            config, self.weight_file, None, char_emb_layer)
-        elmo_model["char_cnn"]['char_emb_layer.weight'] = char_emb_layer.weight
-        self.token_embedder.load_state_dict(elmo_model["char_cnn"])
-
-        self.output_dim = config['lstm']['projection_dim']
-
-        # lstm encoder
-        self.encoder = ElmobiLm(config)
-        self.encoder.load_state_dict(elmo_model["lstm"])
-
-        if cache_word_reprs:
-            if config['char_cnn']['embedding']['dim'] > 0:  # 只有在使用了chars的情况下有用
-                print("Start to generate cache word representations.")
-                batch_size = 320
-                # bos eos
-                word_size = self.words_to_chars_embedding.size(0)
-                num_batches = word_size // batch_size + \
-                              int(word_size % batch_size != 0)
-
-                self.cached_word_embedding = nn.Embedding(word_size,
-                                                          config['lstm']['projection_dim'])
-                with torch.no_grad():
-                    for i in range(num_batches):
-                        words = torch.arange(i * batch_size,
-                                             min((i + 1) * batch_size, word_size)).long()
-                        chars = self.words_to_chars_embedding[words].unsqueeze(1)  # batch_size x 1 x max_chars
-                        word_reprs = self.token_embedder(words.unsqueeze(1),
-                                                         chars).detach()  # batch_size x 1 x config['encoder']['projection_dim']
-                        self.cached_word_embedding.weight.data[words] = word_reprs.squeeze(1)
-
-                    print("Finish generating cached word representations. Going to delete the character encoder.")
-                del self.token_embedder, self.words_to_chars_embedding
-            else:
-                print("There is no need to cache word representations, since no character information is used.")
-
-    def forward(self, words):
-        """
-
-        :param words: batch_size x max_len
-        :return: num_layers x batch_size x max_len x hidden_size
-        """
-        # 扩展<bos>, <eos>
-        batch_size, max_len = words.size()
-        expanded_words = words.new_zeros(batch_size, max_len + 2)  # 因为pad一定为0，
-        seq_len = words.ne(self._pad_index).sum(dim=-1)
-        expanded_words[:, 1:-1] = words
-        expanded_words[:, 0].fill_(self.bos_index)
-        expanded_words[torch.arange(batch_size).to(words), seq_len + 1] = self.eos_index
-        seq_len = seq_len + 2
-        zero_tensor = expanded_words.new_zeros(expanded_words.shape)
-        mask = (expanded_words == zero_tensor).unsqueeze(-1)
-        if hasattr(self, 'cached_word_embedding'):
-            token_embedding = self.cached_word_embedding(expanded_words)
-        else:
-            if hasattr(self, 'words_to_chars_embedding'):
-                chars = self.words_to_chars_embedding[expanded_words]
-            else:
-                chars = None
-            token_embedding = self.token_embedder(expanded_words, chars)  # batch_size x max_len x embed_dim
-
-        encoder_output = self.encoder(token_embedding, seq_len)
-        if encoder_output.size(2) < max_len + 2:
-            num_layers, _, output_len, hidden_size = encoder_output.size()
-            dummy_tensor = encoder_output.new_zeros(num_layers, batch_size,
-                                                    max_len + 2 - output_len, hidden_size)
-            encoder_output = torch.cat((encoder_output, dummy_tensor), 2)
-        sz = encoder_output.size()  # 2, batch_size, max_len, hidden_size
-        token_embedding = token_embedding.masked_fill(mask, 0)
-        token_embedding = torch.cat((token_embedding, token_embedding), dim=2).view(1, sz[1], sz[2], sz[3])
-        encoder_output = torch.cat((token_embedding, encoder_output), dim=0)
-
-        # 删除<eos>, <bos>. 这里没有精确地删除，但应该也不会影响最后的结果了。
-        encoder_output = encoder_output[:, :, 1:-1]
-        return encoder_output
